feat: add communication of size map and place available in a team to AI

+++ server/include/ia/team.h
@@ -16,7 +16,7 @@
 typedef struct s_team {
     list_t *list_ai;
     char *name;
-    int max_place;
+    size_t max_place;
 } team_t;
 
 /**
